# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Video Link Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Folder {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  linkCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface FolderInput {
  /** @minLength 1 */
  name: string;
  description?: string;
}

export interface FolderUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
}

export type BackupLinkStatus = typeof BackupLinkStatus[keyof typeof BackupLinkStatus];


export const BackupLinkStatus = {
  active: 'active',
  expired: 'expired',
  unknown: 'unknown',
} as const;

export interface BackupLink {
  id: number;
  videoLinkId: number;
  /** @nullable */
  label?: string | null;
  url: string;
  priority: number;
  status: BackupLinkStatus;
  /** @nullable */
  lastChecked?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface BackupLinkInput {
  /** @minLength 1 */
  url: string;
  label?: string;
}

export interface BackupLinkUpdate {
  /** @minLength 1 */
  url?: string;
  label?: string;
  priority?: number;
}

export type VideoLinkStatus = typeof VideoLinkStatus[keyof typeof VideoLinkStatus];


export const VideoLinkStatus = {
  active: 'active',
  expired: 'expired',
  checking: 'checking',
  unknown: 'unknown',
} as const;

export interface VideoLink {
  id: number;
  /** @nullable */
  folderId?: number | null;
  /** @nullable */
  folderName?: string | null;
  title: string;
  url: string;
  /** @nullable */
  pageUrl?: string | null;
  /** @nullable */
  refreshedUrl?: string | null;
  /** @nullable */
  activeBackupId?: number | null;
  backupCount: number;
  status: VideoLinkStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  lastChecked?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface VideoLinkInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  url: string;
  pageUrl?: string;
  folderId?: number;
  notes?: string;
}

export type VideoLinkUpdateStatus = typeof VideoLinkUpdateStatus[keyof typeof VideoLinkUpdateStatus];


export const VideoLinkUpdateStatus = {
  active: 'active',
  expired: 'expired',
  checking: 'checking',
  unknown: 'unknown',
} as const;

export interface VideoLinkUpdate {
  /** @minLength 1 */
  title?: string;
  url?: string;
  pageUrl?: string;
  notes?: string;
  status?: VideoLinkUpdateStatus;
}

export interface MoveLinkInput {
  /** @nullable */
  folderId?: number | null;
}

export interface CheckAllResult {
  total: number;
  checked: number;
  active: number;
  expired: number;
  failed: number;
}

export interface Stats {
  totalLinks: number;
  activeLinks: number;
  expiredLinks: number;
  unknownLinks: number;
  totalFolders: number;
  recentlyChecked: number;
}

export type ListLinksParams = {
folderId?: number;
status?: ListLinksStatus;
};

export type ListLinksStatus = typeof ListLinksStatus[keyof typeof ListLinksStatus];


export const ListLinksStatus = {
  active: 'active',
  expired: 'expired',
  checking: 'checking',
  unknown: 'unknown',
} as const;