# lib/api-zod/src/generated/types/backupLink.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Video Link Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { BackupLinkStatus } from './backupLinkStatus';

export interface BackupLink {
  id: number;
  videoLinkId: number;
  /** @nullable */
  label?: string | null;
  url: string;
  priority: number;
  status: BackupLinkStatus;
  /** @nullable */
  lastChecked?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}