# lib/api-zod/src/generated/types/backupLinkUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Video Link Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface BackupLinkUpdate {
  /** @minLength 1 */
  url?: string;
  label?: string;
  priority?: number;
}