# lib/api-zod/src/generated/types/videoLink.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Video Link Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { VideoLinkStatus } from './videoLinkStatus';

export interface VideoLink {
  id: number;
  /** @nullable */
  folderId?: number | null;
  /** @nullable */
  folderName?: string | null;
  title: string;
  url: string;
  /** @nullable */
  pageUrl?: string | null;
  /** @nullable */
  refreshedUrl?: string | null;
  /** @nullable */
  activeBackupId?: number | null;
  backupCount: number;
  status: VideoLinkStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  lastChecked?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}